# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trading Buddy API",
    "description": "A股量化交易数据基础设施",
    "version": "1.1.0"
  },
  "paths": {
    "/api/stocks/list": {
      "get": {
        "tags": [
          "股票"
        ],
        "summary": "Get Stock List",
        "description": "获取股票代码列表（分页），含 `name` 展示名与过滤后总数 `total`。",
        "operationId": "get_stock_list_api_stocks_list_get",
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "市场: sh/sz/bj",
              "title": "Market"
            },
            "description": "市场: sh/sz/bj"
          },
          {
            "name": "industry",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "行业前缀（与 GET /stocks/industry/{industry} 同口径）；未传或空则不过滤",
              "title": "Industry"
            },
            "description": "行业前缀（与 GET /stocks/industry/{industry} 同口径）；未传或空则不过滤"
          },
          {
            "name": "stock_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "类型: common | st | star | growth | beijing；未知值忽略（不过滤）",
              "title": "Stock Type"
            },
            "description": "类型: common | st | star | growth | beijing；未知值忽略（不过滤）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "本页条数（结果按 code 排序后切片）",
              "default": 100,
              "title": "Limit"
            },
            "description": "本页条数（结果按 code 排序后切片）"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "跳过前 offset 条；超过结果集时服务端钳制，响应 `offset` 为实际起点",
              "default": 0,
              "title": "Offset"
            },
            "description": "跳过前 offset 条；超过结果集时服务端钳制，响应 `offset` 为实际起点"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stocks/{code}": {
      "get": {
        "tags": [
          "股票"
        ],
        "summary": "Get Stock",
        "description": "获取股票详情",
        "operationId": "get_stock_api_stocks__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/StockInfo"
                    },
                    {
                      "type": "object",
                      "additionalProperties": true
                    }
                  ],
                  "title": "Response Get Stock Api Stocks  Code  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stocks/industry/{industry}": {
      "get": {
        "tags": [
          "股票"
        ],
        "summary": "Get Stocks By Industry",
        "description": "根据行业获取股票（行业字段**前缀匹配**，如「新能源」可命中「新能源 整车」）。",
        "operationId": "get_stocks_by_industry_api_stocks_industry__industry__get",
        "parameters": [
          {
            "name": "industry",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Industry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockInfo"
                  },
                  "title": "Response Get Stocks By Industry Api Stocks Industry  Industry  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/paper/orders": {
      "get": {
        "tags": [
          "纸交易"
        ],
        "summary": "Paper List Orders",
        "description": "分页查询成交记录（按订单 id 新→旧）。",
        "operationId": "paper_list_orders_api_paper_orders_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "每页条数",
              "default": 50,
              "title": "Limit"
            },
            "description": "每页条数"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "跳过条数",
              "default": 0,
              "title": "Offset"
            },
            "description": "跳过条数"
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 24
                },
                {
                  "type": "null"
                }
              ],
              "description": "可选：只查该标的",
              "title": "Code"
            },
            "description": "可选：只查该标的"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperOrdersListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "纸交易"
        ],
        "summary": "Paper Place Order",
        "operationId": "paper_place_order_api_paper_orders_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperOrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Market": {
        "type": "string",
        "enum": [
          "sh",
          "sz",
          "bj",
          "hk",
          "us"
        ],
        "title": "Market",
        "description": "市场类型"
      },
      "PaperOrderListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "fill_price": {
            "type": "number",
            "title": "Fill Price"
          },
          "fill_amount": {
            "type": "number",
            "title": "Fill Amount"
          },
          "trade_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Date"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "side",
          "quantity",
          "fill_price",
          "fill_amount",
          "trade_date",
          "created_at"
        ],
        "title": "PaperOrderListItem"
      },
      "PaperOrderRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 24,
            "minLength": 4,
            "title": "Code",
            "description": "标的代码，如 sh.600000"
          },
          "side": {
            "type": "string",
            "title": "Side",
            "description": "buy 或 sell"
          },
          "quantity": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 100.0,
            "title": "Quantity",
            "description": "股数：须为 100 的整数倍（A 股一手）"
          }
        },
        "type": "object",
        "required": [
          "code",
          "side",
          "quantity"
        ],
        "title": "PaperOrderRequest"
      },
      "PaperOrderResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "fill_price": {
            "type": "number",
            "title": "Fill Price"
          },
          "fill_amount": {
            "type": "number",
            "title": "Fill Amount"
          },
          "trade_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Date"
          },
          "cash_after": {
            "type": "number",
            "title": "Cash After"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "side",
          "quantity",
          "fill_price",
          "fill_amount",
          "trade_date",
          "cash_after"
        ],
        "title": "PaperOrderResponse"
      },
      "PaperOrdersListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PaperOrderListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaperOrdersListResponse"
      },
...
      "StockInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "股票代码，如 000001.SZ"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "股票名称"
          },
          "ipo_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ipo Date",
            "description": "上市日期"
          },
          "out_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Out Date",
            "description": "退市日期"
          },
          "stock_type": {
            "$ref": "#/components/schemas/StockType",
            "default": "common"
          },
          "market": {
            "$ref": "#/components/schemas/Market",
            "description": "交易市场"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry",
            "description": "所属行业"
          },
          "sector_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Code",
            "description": "板块代码（概念/行业）"
          },
          "is_trading": {
            "type": "boolean",
            "title": "Is Trading",
            "description": "是否交易中（未停牌）",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "market"
        ],
        "title": "StockInfo",
        "description": "股票基本信息"
      },
      "StockListItem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "StockListItem"
      },
      "StockListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StockListItem"
            },
            "type": "array",
            "title": "Items",
            "description": "本页行（按 code 排序后切片）"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "过滤后的总条数（分页前）"
          },
          "limit": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "本页条数上限"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "本页实际起始跳过条数（请求 `offset` 超界时钳到最后一页起点）"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "StockListResponse",
        "description": "分页列表体；`/docs` 中展示示例 JSON。",
        "examples": [
          {
            "items": [
              {
                "code": "sh.600000",
                "name": "浦发银行",
                "status": "ok"
              },
              {
                "code": "sz.000001",
                "name": "平安银行",
                "status": "ok"
              }
            ],
            "limit": 100,
            "offset": 0,
            "total": 4281
          }
        ]
      },
      "StockType": {
        "type": "string",
        "enum": [
          "common",
          "st",
          "star",
          "growth",
          "beijing"
        ],
        "title": "StockType",
        "description": "股票类型"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}